requirements.txt
lightning==2.5.1.post0
matplotlib==3.10.6
numpy==2.3.2
overrides==7.7.0
Pillow==11.3.0
scipy==1.16.1
torch==2.5.1+cu124
torchcvnn==0.5.0
torchmetrics==1.5.0
torchvision==0.20.1+cu124
tqdm==4.66.5
tensorboard==2.19.0
tensorboardX==2.6.2.2
